Move helpers from CiviUnitTestCase to GenericAssertionsTrait
* `assertType(...)`
* `assertTreeEquals(...)`
* `assertApproxEquals(...)`
* `assertAttributesEquals(...)`
* `assertArrayKeyExists(...)`
* `assertArrayValueNotNull(...)`
* `assertArrayValuesEqual(...)`
